Update cricket.py
No need of creating empty strings variable, variablewill be automatically created when you will take the input from the user.

There is no need of keeping the terminal awake as everyone runs it either from their IDE or VS Code or sometimes directly terminal, as double clicking the file and running it is a bad practice. But still if you want to keep the terminal awake for people who double click the file I have kept the terminal awake for 60 secs you can increase or decrease it.

I hope you will find this PR useful and merge it.
Thanks.